Add (basic) publish method to Rakefile
Acknowledging David Ensinger in the Readme

+++ Rakefile
@@ -11,9 +11,21 @@ task :clean do
 end
 
 # "--auto" regenerates the site automatically when files are changed
-desc 'Runs Jekyll server locally on port 4000'
+desc 'Run the jekyll server on port 4000'
 task :dev do
   system('jekyll --server --auto')
+end
+
+# TODO: I should be able to not type the post date, i.e. not have 
+# to type `rake publish[YYYY-MM-DD-the-post-title]`
+desc 'Publish a post from _drafts'
+task :publish, :title do |command, args|
+  filename = args.title.split( /_drafts\/|_posts\// ).last
+  raise RuntimeError.new('no draft specified!') if filename.nil? || filename.empty?
+  puts "Removing _posts/#{filename}"
+  system("rm _posts/#{filename}")
+  puts "Moving #{filename} from _drafts/ to _posts/"
+  system("mv _drafts/#{filename} _posts/#{filename}")
 end
 
 # Note: all my posts are in the "posts" category, but I may add additional 
@@ -22,8 +34,9 @@ end
 desc "Given a title as an argument, create a new post file"
 task :post, :title do |le_task, args|
   filename = "#{Time.now.strftime('%Y-%m-%d')}-#{args.title.gsub(/\s/, '_').downcase}.markdown"
-  path = File.join("_posts", filename)
+  path = File.join("_drafts", filename)
   if File.exist? path; raise RuntimeError.new("Won't clobber #{path}"); end
+  raise RuntimeError.new("Won't clobber #{path}") if File.exist?(path)
   File.open(path, 'w') do |file|
     file.write <<-EOS
 ---
@@ -34,5 +47,6 @@ date: #{Time.now.strftime('%Y-%m-%d %k:%M:%S')}
 ---
 EOS
     end
-    puts "Template post created at: #{path} \nGo forth and blog!"
+  system("ln #{path} _posts/#{filename}")
+  puts "Template post created at: #{path} \nGo forth and blog!"
 end
